revise caching concept
Previously, files would be written to disk and the path would be cached. It could not be guaranteed that everything would always be deleted from disk again - this lead to accumulation of files.
Now, most operations are done in-memory (using io.BytesIO objects) and only written to disk when necessary for ffmpeg. This allows in-memory caching of the io.BytesIO objects, ensuring cleanup. The OS swapping is
now responsible for persisting the memory to disk in a low-memory situation.