add title property to VideoPlotter (incl. automatic one); speed up filling of database
By batching the transactions, I can get a speedup of more than 800% (15.63s/it to 1.92s/it). Which is great because it now doesn't take more than a week to run but just slightly more than a day.
The addition of the timestamp to the DB is required for other applications anyway - might be better to make it an index at some point, though.