plotting: swap X/Y coordinates and rotate images correctly
Not only does this account for the up-to-date bb_binary version (where X and Y are swapped/corrected), but in combination with the flipped axes, the final images are now just as they were in reality.
That means: up is up and left is left. The camera rotations are now corrected.

Also, the minimum path alpha now also scales with the default alpha. Thought I had committed that already.