allow FramePlotter to cache all images in the respective frame container
VideoPlotter already did something similar: every video file had only to be touched once.
However, when calling FramePlotter with a sequence of images that belong to the same video, ffmpeg would be used to extract single frames for every call and touch the file N times.
This was rather slow. Now you can request the backend to decode all frames in the video when you first request a frame. On successive calls, these cached images will be used and ffmpeg will not be invoked again.